Rename package from com.netflix.spinnaker.kayenta to com.netflix.kayenta.
