+++ build.gradle
@@ -29,7 +29,7 @@ buildscript {
 }
 
 allprojects {
-  group = "com.netflix.spinnaker.kayenta"
+  group = "com.netflix.kayenta"
   apply plugin: 'spinnaker.project'
   apply plugin: 'groovy'
