+++ kayenta-core/src/main/java/com/netflix/kayenta/config/KayentaConfiguration.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2017 Google, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License")
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.netflix.kayenta.config;
+
+import com.netflix.kayenta.metrics.MapBackedMetricsServiceRepository;
+import com.netflix.kayenta.metrics.MetricsServiceRepository;
+import com.netflix.kayenta.security.AccountCredentialsRepository;
+import com.netflix.kayenta.security.MapBackedAccountCredentialsRepository;
+import com.netflix.kayenta.storage.MapBackedStorageServiceRepository;
+import com.netflix.kayenta.storage.StorageServiceRepository;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+
+public class KayentaConfiguration {
+
+  @Bean
+  @ConditionalOnMissingBean(AccountCredentialsRepository.class)
+  AccountCredentialsRepository accountCredentialsRepository() {
+    return new MapBackedAccountCredentialsRepository();
+  }
+
+  @Bean
+  @ConditionalOnMissingBean(MetricsServiceRepository.class)
+  MetricsServiceRepository metricsServiceRepository() {
+    return new MapBackedMetricsServiceRepository();
+  }
+
+  @Bean
+  @ConditionalOnMissingBean(StorageServiceRepository.class)
+  StorageServiceRepository storageServiceRepository() {
+    return new MapBackedStorageServiceRepository();
+  }
+}
