--- kayenta-core/src/main/java/com/netflix/spinnaker/kayenta/config/KayentaConfiguration.java
@@ -1,47 +1,0 @@
-/*
- * Copyright 2017 Google, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License")
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netflix.spinnaker.kayenta.config;
-
-import com.netflix.spinnaker.kayenta.metrics.MapBackedMetricsServiceRepository;
-import com.netflix.spinnaker.kayenta.metrics.MetricsServiceRepository;
-import com.netflix.spinnaker.kayenta.security.AccountCredentialsRepository;
-import com.netflix.spinnaker.kayenta.security.MapBackedAccountCredentialsRepository;
-import com.netflix.spinnaker.kayenta.storage.MapBackedStorageServiceRepository;
-import com.netflix.spinnaker.kayenta.storage.StorageServiceRepository;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.context.annotation.Bean;
-
-public class KayentaConfiguration {
-
-  @Bean
-  @ConditionalOnMissingBean(AccountCredentialsRepository.class)
-  AccountCredentialsRepository accountCredentialsRepository() {
-    return new MapBackedAccountCredentialsRepository();
-  }
-
-  @Bean
-  @ConditionalOnMissingBean(MetricsServiceRepository.class)
-  MetricsServiceRepository metricsServiceRepository() {
-    return new MapBackedMetricsServiceRepository();
-  }
-
-  @Bean
-  @ConditionalOnMissingBean(StorageServiceRepository.class)
-  StorageServiceRepository storageServiceRepository() {
-    return new MapBackedStorageServiceRepository();
-  }
-}
