--- kayenta-core/src/main/java/com/netflix/spinnaker/kayenta/security/AccountCredentialsRepository.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2017 Google, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License")
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netflix.spinnaker.kayenta.security;
-
-import java.util.Optional;
-import java.util.Set;
-
-public interface AccountCredentialsRepository {
-  Optional<AccountCredentials> getOne(String accountName);
-
-  Optional<AccountCredentials> getOne(AccountCredentials.Type credentialsType);
-
-  Set<? extends AccountCredentials> getAll();
-
-  AccountCredentials save(String name, AccountCredentials credentials);
-}
