--- kayenta-google/src/main/java/com/netflix/spinnaker/kayenta/google/config/GoogleConfigurationProperties.java
@@ -1,28 +1,0 @@
-/*
- * Copyright 2017 Google, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License")
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netflix.spinnaker.kayenta.google.config;
-
-import lombok.Getter;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class GoogleConfigurationProperties {
-
-  @Getter
-  private List<GoogleManagedAccount> accounts = new ArrayList<>();
-}
